+++ libs/render_delegate/reader.h
@@ -56,4 +56,5 @@ private:
     GfVec2f _shutter;
     HdTaskSharedPtrVector _tasks;
     HdTaskContext _taskContext;
+    std::vector<AtNode*> _nodes;
 };